requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
gunicorn==20.0.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
